Delete Plaso.mkape
Eric, would you mind implementing this PR. This module I wrote was just an experiment and it provides no value in it's current form. If anyone thinks they would actually use this in the field or the lab, I will rewrite using mvars which will allow for a much more targeted time line.  -m