initial Commit
Initial commit.  The functionality of this KAPE Target is mostly replicated in the !BasicCollection target.  The difference is that no compound targets are used to make this target self-documenting for the SANS FOR500 student.  We want them to be able to look at one file to see all the actual files, and not just the category, we collect in a triage image.