Process Creation Details via Get-CimInstance_ProcessList
The current module PWSH-Get-ProcessList.mkape is very slow in receiving details on a live system. During tests on a remote live response session this ended in a Timeout and closing of the session. This is a proposal to use the Get-CimInstance Win32_Process approach instead which only takes between 10 - 20 secs as opposed to 20 - 30 Mins.