Add required folder to BrowsingHistoryView command
We use sourceDirectory to be flexible if source drive is C or whatever, but add only the required part "users" to the path, so it works directly when invoking e.g. `kape.exe -msource D:\C` or `kape.exe -msource D:\E` or whatever. This is now a combination from the previous too static %sourceDirectory%\C\Users\ and the last one using only %sourceDirectory% which needed always a D:\C\Users explicitly in the -msource.